Move SSO account into account package
