+++ backend/app/usecase/sso/sso.go
@@ -6,16 +6,15 @@ import (
 	"github.com/short-d/short/backend/app/entity"
 	"github.com/short-d/short/backend/app/usecase/account"
 	"github.com/short-d/short/backend/app/usecase/authenticator"
-	"github.com/short-d/short/backend/app/usecase/external"
 )
 
 // SingleSignOn enables sign in through external identity providers, such as
 // Github, Facebook, and Google.
 type SingleSignOn struct {
-	identityProvider  IdentityProvider
-	ssoAccountService external.SSOAccount
-	accountProvider   account.Provider
-	authenticator     authenticator.Authenticator
+	identityProvider IdentityProvider
+	ssoAccount       account.SSOAccount
+	accountProvider  account.Provider
+	authenticator    authenticator.Authenticator
 }
 
 // SignIn generates access token for a user using authorization code obtained
@@ -30,7 +29,7 @@ func (o SingleSignOn) SignIn(authorizationCode string) (string, error) {
 		return "", err
 	}
 
-	ssoUser, err := o.ssoAccountService.GetSingleSignOnUser(accessToken)
+	ssoUser, err := o.ssoAccount.GetSingleSignOnUser(accessToken)
 	if err != nil {
 		return "", err
 	}
@@ -65,14 +64,14 @@ func (o SingleSignOn) SignIn(authorizationCode string) (string, error) {
 // identity provider.
 func NewSingleSignOn(
 	identityProvider IdentityProvider,
-	ssoAccountService external.SSOAccount,
+	ssoAccount account.SSOAccount,
 	accountProvider account.Provider,
 	authenticator authenticator.Authenticator,
 ) SingleSignOn {
 	return SingleSignOn{
-		identityProvider:  identityProvider,
-		ssoAccountService: ssoAccountService,
-		accountProvider:   accountProvider,
-		authenticator:     authenticator,
+		identityProvider: identityProvider,
+		ssoAccount:       ssoAccount,
+		accountProvider:  accountProvider,
+		authenticator:    authenticator,
 	}
 }
