+++ backend/app/usecase/sso/sso_test.go
@@ -12,7 +12,6 @@ import (
 	"github.com/short-d/short/backend/app/entity"
 	"github.com/short-d/short/backend/app/usecase/account"
 	"github.com/short-d/short/backend/app/usecase/authenticator"
-	"github.com/short-d/short/backend/app/usecase/external"
 	"github.com/short-d/short/backend/app/usecase/repository"
 )
 
@@ -61,7 +60,7 @@ func TestSingleSignOn_SignIn(t *testing.T) {
 			t.Parallel()
 
 			identityProvider := NewIdentityProviderFake("http://localhost/sign-in", "")
-			profileService := external.NewSSOAccountFake(testCase.ssoUser)
+			profileService := account.NewSSOAccountFake(testCase.ssoUser)
 			fakeUserRepo := repository.NewUserFake(testCase.users)
 			tm := timer.NewStub(time.Now())
 			accountProvider := account.NewProvider(&fakeUserRepo, tm)
